requirements.txt
psutil==5.7.2
pure-sasl==0.6.2
pydle==0.9.4
python-pidfile==3.0.0
